2/27
I figured out the hunger situation. It took a long time to figure out
but now I need to add all the other food and into the list of what will
satisfy the player. Then now I have to create the NPC who will be apart
of another puzzle to later win the game.